fix(tools): cast rustix stat types for macOS (st_mode u16→u32, st_dev i32→u64)
rustix::fs::statat returns st_mode as u16 and st_dev as i32 on macOS,
but std::fs::Permissions::from_mode expects u32 and std::fs::Metadata::dev()
returns u64. Add explicit casts to bridge the type gap.

+++ src/tools.rs
@@ -4210,7 +4210,7 @@ where
         let original_permissions = expected_target_identity.map(|(_, _, mode)| {
             use std::os::unix::fs::PermissionsExt as _;
 
-            std::fs::Permissions::from_mode(mode)
+            std::fs::Permissions::from_mode(mode as u32)
         });
 
         // The pathname stored by `NamedTempFile` is not descriptor-relative.
@@ -4307,8 +4307,8 @@ where
                 let source_file = std::fs::File::from(source_descriptor);
                 let source_metadata = source_file.metadata()?;
                 if !source_metadata.is_file()
-                    || source_metadata.dev() != expected_dev
-                    || source_metadata.ino() != expected_ino
+                    || source_metadata.dev() != expected_dev as u64
+                    || source_metadata.ino() != expected_ino as u64
                 {
                     return Err(std::io::Error::other(
                         "file changed since it was read; re-read it and retry the edit",
